+++ src/main/java/com/lokoko/domain/campaign/application/mapper/CampaignMapper.java
@@ -2,7 +2,7 @@ package com.lokoko.domain.campaign.application.mapper;
 
 import com.lokoko.domain.campaign.api.dto.response.CampaignParticipatedResponse;
 import com.lokoko.domain.campaign.domain.entity.Campaign;
-import com.lokoko.domain.campaign.domain.entity.CreatorCampaign;
+import com.lokoko.domain.creatorCampaign.domain.entity.CreatorCampaign;
 import org.springframework.stereotype.Component;
 
 @Component
