+++ src/main/java/com/lokoko/domain/campaign/application/service/CampaignGetService.java
@@ -3,11 +3,11 @@ package com.lokoko.domain.campaign.application.service;
 import com.lokoko.domain.campaign.api.dto.response.CampaignDetailResponse;
 import com.lokoko.domain.campaign.api.dto.response.CampaignImageResponse;
 import com.lokoko.domain.campaign.domain.entity.Campaign;
-import com.lokoko.domain.campaign.domain.entity.CreatorCampaign;
 import com.lokoko.domain.campaign.domain.entity.enums.CampaignDetailPageStatus;
 import com.lokoko.domain.campaign.domain.repository.CampaignRepository;
-import com.lokoko.domain.campaign.domain.repository.CreatorCampaignRepository;
 import com.lokoko.domain.campaign.exception.CampaignNotFoundException;
+import com.lokoko.domain.creatorCampaign.domain.entity.CreatorCampaign;
+import com.lokoko.domain.creatorCampaign.domain.repository.CreatorCampaignRepository;
 import com.lokoko.domain.image.domain.repository.CampaignImageRepository;
 import java.util.List;
 import java.util.Optional;
