+++ src/main/java/com/lokoko/domain/campaignReview/application/usecase/CampaignReviewUsecase.java
@@ -1,7 +1,8 @@
 package com.lokoko.domain.campaignReview.application.usecase;
 
 
-import com.lokoko.domain.campaign.domain.entity.CreatorCampaign;
+import com.lokoko.domain.campaign.api.dto.response.CampaignParticipatedResponse;
+import com.lokoko.domain.campaign.application.mapper.CampaignMapper;
 import com.lokoko.domain.campaignReview.api.dto.request.FirstReviewUploadRequest;
 import com.lokoko.domain.campaignReview.api.dto.request.SecondReviewUploadRequest;
 import com.lokoko.domain.campaignReview.api.dto.response.ReviewUploadResponse;
@@ -15,6 +16,9 @@ import com.lokoko.domain.campaignReview.exception.FirstReviewNotFoundException;
 import com.lokoko.domain.campaignReview.exception.MismatchedContentTypeException;
 import com.lokoko.domain.creator.application.service.CreatorGetService;
 import com.lokoko.domain.creator.domain.entity.Creator;
+import com.lokoko.domain.creatorCampaign.application.service.CreatorCampaignGetService;
+import com.lokoko.domain.creatorCampaign.domain.entity.CreatorCampaign;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -25,11 +29,13 @@ public class CampaignReviewUsecase {
 
     private final CreatorGetService creatorGetService;
     private final CampaignReviewGetService campaignReviewGetService;
+    private final CreatorCampaignGetService creatorCampaignGetService;
 
     private final CampaignReviewSaveService campaignReviewSaveService;
     private final CreatorCampaignUpdateService creatorCampaignUpdateService;
 
     private final CampaignReviewMapper campaignReviewMapper;
+    private final CampaignMapper campaignMapper;
 
     @Transactional
     public ReviewUploadResponse uploadFirst(Long userId, Long campaignId, FirstReviewUploadRequest request) {
@@ -74,5 +80,14 @@ public class CampaignReviewUsecase {
         creatorCampaignUpdateService.refreshParticipationStatus(participation.getId());
 
         return campaignReviewMapper.toUploadResponse(saved);
+    }
+
+    @Transactional(readOnly = true)
+    public List<CampaignParticipatedResponse> getMyReviewableCampaigns(Long userId) {
+        Creator creator = creatorGetService.findByUserId(userId);
+        List<CreatorCampaign> eligible = creatorCampaignGetService.findReviewAble(creator.getId());
+        return eligible.stream()
+                .map(campaignMapper::toCampaignParticipationResponse)
+                .toList();
     }
 }
