+++ src/main/java/com/lokoko/domain/creator/application/service/CreatorUpdateService.java
@@ -1,12 +1,12 @@
 package com.lokoko.domain.creator.application.service;
 
 
-import com.lokoko.domain.campaign.domain.entity.CreatorCampaign;
 import com.lokoko.domain.campaignReview.application.service.CreatorCampaignUpdateService;
 import com.lokoko.domain.creator.api.dto.request.CreatorInfoUpdateRequest;
 import com.lokoko.domain.creator.api.dto.request.CreatorMyPageUpdateRequest;
 import com.lokoko.domain.creator.domain.entity.Creator;
 import com.lokoko.domain.creator.exception.CreatorIdAlreadyExistsException;
+import com.lokoko.domain.creatorCampaign.domain.entity.CreatorCampaign;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
